"Update various functions, add constants and adjust data types"
This commit includes a variety of changes in multiple files. It includes the addition of new constants in the Constant.ts, including a significant constant for 'TWO_WEEKS_IN_SECONDS'. Also, there are amendments to the CamelotUniswapV3Pool and CamelotUniswapV3Vault event handler and the addition of new methods to PlatformUtils.ts. Changes in the subgraph.yaml file involve an upgrade to the 'specVersion' and 'apiVersion'. UserBalance.ts  shows alterations in several functions and the addition of the 'updateVaultUsers' function. New types and changes to several methods were made in the schema file.